rw-14: refactor upload to custom component
